fix(sdk): remove dead vrm.ts and relocate repeat/TimelineKeyframe
After the persona-first migration, the 833-line vrm.ts was almost
entirely unused. Move the `repeat` namespace to persona.ts and
`TimelineKeyframe` interface to speech.ts, then delete vrm.ts.
