+++ packages/sdk/src/index.ts
@@ -48,7 +48,6 @@ export * from "./preferences";
 export * from "./math";
 export * from "./settings";
 export * from "./shadowPanel";
-export * from "./vrm";
 export * from "./persona";
 export * from "./speech";
 export * from "./webviews";
